+++ extend/bajo/hook/dobo@before-find-one-record.js
@@ -1,8 +1,0 @@
-import { handler } from './dobo@before-find-record.js'
-
-const doboBeforeFindOneRecord = {
-  level: 1000,
-  handler
-}
-
-export default doboBeforeFindOneRecord