+++ wiki/CHANGES.md
@@ -1,5 +1,11 @@
 # Changes
 
+## 2026-03-11
+
+- [2.8.0] Add ```createNewSite()``` and ```applet.crateNewSite```
+- [2.8.0] Add ```removeSite()``` and ```applet.removeSite```
+- [2.8.0] Set ```site.json```, ```user.json```, ```team.json```, ```team-user.json``` first fixture as immutable row
+
 ## 2026-03-08
 
 - [2.7.3] Bug fix on ```mergeTeam()```